# feat: account for instance type requirements for daemonset scheduling