@saurav-agarwalla 2025-11-06
@logyball made some performance optimizations, added the test and also ensured that the (now) minor performance penalty is only for when customers use daemon set pods.

New results are below and I think they look reasonable:
```
➜  ~ benchstat /tmp/old /tmp/new
goos: darwin
goarch: arm64
pkg: sigs.k8s.io/karpenter/pkg/controllers/provisioning/scheduling
cpu: Apple M3 Pro
                      │   /tmp/old   │               /tmp/new                │
                      │    sec/op    │    sec/op      vs base                │
Scheduling1-12          164.3µ ±  8%    190.5µ ±  4%  +15.94% (p=0.000 n=10)
Scheduling50-12         37.51m ±  7%    40.54m ± 15%        ~ (p=0.143 n=10)
Scheduling100-12        76.69m ±  9%    79.59m ± 10%        ~ (p=0.393 n=10)
Scheduling500-12        388.3m ± 27%    404.6m ±  7%        ~ (p=0.684 n=10)
Scheduling1000-12       776.8m ± 13%    898.3m ± 22%        ~ (p=0.075 n=10)
Scheduling2000-12        1.540 ± 12%     1.726 ± 24%  +12.06% (p=0.001 n=10)
Scheduling5000-12        3.738 ± 11%     4.197 ± 12%  +12.30% (p=0.001 n=10)
Scheduling10000-12       8.113 ±  9%    10.153 ± 17%  +25.15% (p=0.000 n=10)
Scheduling20000-12       20.19 ± 25%     22.03 ± 11%        ~ (p=0.105 n=10)
RespectPreferences-12    11.66 ± 15%     10.85 ±  5%   -6.95% (p=0.000 n=10)
IgnorePreferences-12    649.5m ±  6%   1006.5m ±  8%  +54.96% (p=0.000 n=10)
geomean                 534.6m          604.7m        +13.12%

                      │   /tmp/old   │               /tmp/new                │
                      │    nodes     │    nodes      vs base                 │
Scheduling1-12           1.000 ±  0%    1.000 ±  0%       ~ (p=1.000 n=10) ¹
Scheduling50-12          10.00 ±  0%    10.00 ±  0%       ~ (p=1.000 n=10) ¹
Scheduling100-12         20.00 ±  0%    20.00 ±  0%       ~ (p=1.000 n=10) ¹
Scheduling500-12         100.0 ±  0%    100.0 ±  0%       ~ (p=1.000 n=10) ¹
Scheduling1000-12        200.0 ±  0%    200.0 ±  0%       ~ (p=1.000 n=10) ¹
Scheduling2000-12        400.0 ±  0%    400.0 ±  0%       ~ (p=1.000 n=10) ¹
Scheduling5000-12       1.000k ±  0%   1.000k ±  0%       ~ (p=1.000 n=10) ¹
Scheduling10000-12      2.000k ±  0%   2.000k ±  0%       ~ (p=1.000 n=10) ¹
Scheduling20000-12      4.000k ±  0%   4.000k ±  0%       ~ (p=1.000 n=10) ¹
RespectPreferences-12   4.000k ±  0%   4.000k ±  0%       ~ (p=1.000 n=10) ¹
IgnorePreferences-12    11.000 ± 45%    8.500 ± 88%       ~ (p=0.444 n=10)
geomean                  144.3          140.9        -2.32%
¹ all samples are equal

                      │  /tmp/old   │               /tmp/new               │
                      │    pods     │    pods      vs base                 │
Scheduling1-12           1.000 ± 0%    1.000 ± 0%       ~ (p=1.000 n=10) ¹
Scheduling50-12          50.00 ± 0%    50.00 ± 0%       ~ (p=1.000 n=10) ¹
Scheduling100-12         100.0 ± 0%    100.0 ± 0%       ~ (p=1.000 n=10) ¹
Scheduling500-12         500.0 ± 0%    500.0 ± 0%       ~ (p=1.000 n=10) ¹
Scheduling1000-12       1.000k ± 0%   1.000k ± 0%       ~ (p=1.000 n=10) ¹
Scheduling2000-12       2.000k ± 0%   2.000k ± 0%       ~ (p=1.000 n=10) ¹
Scheduling5000-12       5.000k ± 0%   5.000k ± 0%       ~ (p=1.000 n=10) ¹
Scheduling10000-12      10.00k ± 0%   10.00k ± 0%       ~ (p=1.000 n=10) ¹
Scheduling20000-12      20.00k ± 0%   20.00k ± 0%       ~ (p=1.000 n=10) ¹
RespectPreferences-12   4.000k ± 0%   4.000k ± 0%       ~ (p=1.000 n=10) ¹
IgnorePreferences-12    4.000k ± 0%   4.000k ± 0%       ~ (p=1.000 n=10) ¹
geomean                  794.8         794.8       +0.00%
¹ all samples are equal

                      │   /tmp/old   │               /tmp/new               │
                      │   pods/sec   │   pods/sec    vs base                │
Scheduling1-12          6.086k ±  8%   5.250k ±  4%  -13.74% (p=0.000 n=10)
Scheduling50-12         1.333k ±  7%   1.234k ± 13%        ~ (p=0.138 n=10)
Scheduling100-12        1.304k ±  9%   1.256k ± 11%        ~ (p=0.393 n=10)
Scheduling500-12        1.290k ± 21%   1.236k ±  7%        ~ (p=0.644 n=10)
Scheduling1000-12       1.288k ± 11%   1.113k ± 18%        ~ (p=0.072 n=10)
Scheduling2000-12       1.298k ± 14%   1.159k ± 20%  -10.74% (p=0.001 n=10)
Scheduling5000-12       1.338k ± 10%   1.192k ± 11%  -10.95% (p=0.001 n=10)
Scheduling10000-12      1233.0 ±  8%    985.0 ± 14%  -20.12% (p=0.000 n=10)
Scheduling20000-12       990.3 ± 20%    908.6 ± 10%        ~ (p=0.105 n=10)
RespectPreferences-12    343.1 ± 13%    368.7 ±  5%   +7.48% (p=0.000 n=10)
IgnorePreferences-12    6.159k ±  6%   3.974k ±  8%  -35.47% (p=0.000 n=10)
geomean                 1.487k         1.315k        -11.60%
```